pyproject.toml
[project]
name = "agent-one"
version = "0.1.0"
description = "Building or simulating a crude version of Claude Code backed by Google's Gemini API!"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "python-dotenv==1.1.0",
    "ruff>=0.15.1",
]
